Bug, In Session::_revalidateTopology(), iterate over an invalid vector!
* Bug: In Anabatic::Session::_revalidateTopology(), when iterating over
    _segmentInvalidateds, the vector can be modified. If it leads to a
    reallocation we end up on invalid iterators (using freed memory so
    sometimes with overwritten contents). Now, iterate with an index
    which warranty that we get a valid item at each iteration of the
    loop. And, of course, the vector is ensured to not shrink...
* Change: In Anabatic::Session::Session(), reserve (pre-allocate) at least
    1024 elements for all vectors. Mostly prevent the above bug and
    avoid constant reallocation.

+++ anabatic/src/AnabaticEngine.cpp
@@ -1515,7 +1515,7 @@ namespace Anabatic {
         Edge* edge = neighbor->getWestEdge();
         if (not edge) break;
 
-        if (edge->getReservedCapacity() < maxTermSat)
+        if (edge->getReservedCapacity() < (uint32_t)maxTermSat)
           edge->reserveCapacity( maxTermSat - edge->getReservedCapacity() );
         neighbor = neighbor->getWest();
       }
@@ -1524,7 +1524,7 @@ namespace Anabatic {
         Edge* edge = neighbor->getEastEdge();
         if (not edge) break;
 
-        if (edge->getReservedCapacity() < maxTermSat)
+        if (edge->getReservedCapacity() < (uint32_t)maxTermSat)
           edge->reserveCapacity( maxTermSat - edge->getReservedCapacity() );
         neighbor = neighbor->getEast();
       }
